made dialog bigger for midumori
